[risk] Rename core to risk
