+++ projects/ores.client/CMakeLists.txt
@@ -65,7 +65,7 @@ target_include_directories(${exe_target_name}
 
 target_link_libraries(${exe_target_name} PRIVATE
     ${lib_target_name}
-    ores.core.lib
+    ores.risk.lib
     ores.utility.lib
     Boost::boost
     Boost::date_time
