+++ projects/ores.risk.test/CMakeLists.txt
@@ -15,40 +15,41 @@
 # this program; if not, write to the Free Software Foundation, Inc., 51 Franklin
 # Street, Fifth Floor, Boston, MA 02110-1301, USA.
 #
-set(name "ores.core")
-set(lib_binary_name ${name})
+set(name "ores.risk")
 set(lib_target_name ${name}.lib)
+set(tests_binary_name ${name}.test)
+set(tests_target_name ${name}.test)
 
 set(files "")
 file(GLOB_RECURSE files RELATIVE
     "${CMAKE_CURRENT_SOURCE_DIR}/"
     "${CMAKE_CURRENT_SOURCE_DIR}/*.cpp")
 
-set(lib_files ${files})
-add_library(${lib_target_name} ${lib_files})
-set_target_properties(${lib_target_name} PROPERTIES
-    OUTPUT_NAME ${lib_binary_name})
-set_target_properties(${lib_target_name}
-    PROPERTIES VERSION ${DOGEN_VERSION} SOVERSION ${DOGEN_VERSION_MAJOR})
+add_executable(${tests_target_name} ${files})
 
-target_include_directories(${lib_target_name} PUBLIC
-    ${CMAKE_SOURCE_DIR}/projects/
-    ${RAPIDXML_NS_INCLUDE_DIRS}
-)
+set_target_properties(${tests_target_name}
+    PROPERTIES OUTPUT_NAME ${tests_binary_name})
 
-target_link_libraries(${lib_target_name}
-    PRIVATE
-        ores.utility.lib
-        libpqxx::pqxx
-        Boost::boost
-        Boost::program_options
-        sqlgen::sqlgen
-    PUBLIC
-        reflectcpp::reflectcpp
-        PostgreSQL::PostgreSQL
-    )
+target_include_directories(${tests_binary_name}
+    PRIVATE ${CMAKE_SOURCE_DIR}/projects/${name}/include
+    PRIVATE ${CMAKE_SOURCE_DIR}/projects/)
 
-install(TARGETS ${lib_target_name}
-    ARCHIVE DESTINATION lib
-    LIBRARY DESTINATION lib
-    COMPONENT libraries)
+target_link_libraries(${tests_target_name} PRIVATE
+    ${lib_target_name}
+    ores.utility.lib
+    Boost::unit_test_framework
+    ${CMAKE_THREAD_LIBS_INIT})
+
+set(command $<TARGET_FILE:${tests_binary_name}> ${boost_test_parameters})
+
+add_custom_target(run_${tests_target_name}
+    COMMENT "Running ${tests_target_name}" VERBATIM
+    COMMAND ${command}
+    WORKING_DIRECTORY ${stage_bin_dir}
+    DEPENDS ${tests_target_name})
+
+add_dependencies(run_all_tests run_${tests_target_name})
+
+add_boost_tests(${tests_binary_name} ${files})
+
+install(TARGETS ${tests_target_name} RUNTIME DESTINATION bin COMPONENT tests)
