+++ projects/ores.risk.test/main.cpp
@@ -1,6 +1,6 @@
 /* -*- mode: c++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
  *
- * Copyright (C) 2025 Marco Craveiro <marco.craveiro@gmail.com>
+ * Copyright (C) 2012-2015 Marco Craveiro <marco.craveiro@gmail.com>
  *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
@@ -17,24 +17,9 @@
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  *
  */
-#ifndef ORES_CORE_RISK_XML_CURRENCY_CONFIG_SERIALISER_HPP
-#define ORES_CORE_RISK_XML_CURRENCY_CONFIG_SERIALISER_HPP
+#define BOOST_TEST_MODULE ores.risk.test
+#include <boost/test/unit_test.hpp>
+#include "ores.utility/test/fixture.hpp"
 
-#if defined(_MSC_VER) && (_MSC_VER >= 1200)
-#pragma once
-#endif
-
-#include <string>
-#include "ores.core/risk/types/currency_config.hpp"
-
-namespace ores::core::risk::xml {
-
-class currency_config_serialiser {
-public:
-    std::string serialise(const types::currency_config& cfg);
-    types::currency_config deserialise(std::string s);
-};
-
-}
-
-#endif
+using namespace ores::utility::test;
+BOOST_GLOBAL_FIXTURE(exception_fixture);
