+++ projects/ores.risk/CMakeLists.txt
@@ -15,10 +15,7 @@
 # this program; if not, write to the Free Software Foundation, Inc., 51 Franklin
 # Street, Fifth Floor, Boston, MA 02110-1301, USA.
 #
-configure_file("${CMAKE_CURRENT_SOURCE_DIR}/generate_certs.sh"
-    "${CMAKE_BINARY_DIR}/projects/ores.service/generate_certs.sh" COPYONLY)
-
-set(name "ores.service")
+set(name "ores.risk")
 set(lib_binary_name ${name})
 set(lib_target_name ${name}.lib)
 
@@ -28,53 +25,30 @@ file(GLOB_RECURSE files RELATIVE
     "${CMAKE_CURRENT_SOURCE_DIR}/*.cpp")
 
 set(lib_files ${files})
-list(FILTER lib_files EXCLUDE REGEX "main.cpp")
-
 add_library(${lib_target_name} ${lib_files})
 set_target_properties(${lib_target_name} PROPERTIES
     OUTPUT_NAME ${lib_binary_name})
 set_target_properties(${lib_target_name}
     PROPERTIES VERSION ${DOGEN_VERSION} SOVERSION ${DOGEN_VERSION_MAJOR})
 
-target_include_directories(${lib_target_name}
-    PUBLIC ${CMAKE_SOURCE_DIR}/projects/
-    PRIVATE ${CMAKE_CURRENT_BINARY_DIR}
+target_include_directories(${lib_target_name} PUBLIC
+    ${CMAKE_SOURCE_DIR}/projects/
+    ${RAPIDXML_NS_INCLUDE_DIRS}
 )
 
-target_link_libraries(${lib_target_name} PRIVATE
-    OpenSSL::SSL
-    OpenSSL::Crypto
-    Boost::boost
-    Boost::cobalt
-    Boost::program_options)
+target_link_libraries(${lib_target_name}
+    PRIVATE
+        ores.utility.lib
+        libpqxx::pqxx
+        Boost::boost
+        Boost::program_options
+        sqlgen::sqlgen
+    PUBLIC
+        reflectcpp::reflectcpp
+        PostgreSQL::PostgreSQL
+    )
 
 install(TARGETS ${lib_target_name}
     ARCHIVE DESTINATION lib
     LIBRARY DESTINATION lib
     COMPONENT libraries)
-
-set(exe_binary_name ${name})
-set(exe_target_name ${name}.exe)
-
-set(exe_files ${files})
-list(FILTER exe_files INCLUDE REGEX "main.cpp")
-
-add_executable(${exe_target_name} ${exe_files})
-set_target_properties(${exe_target_name} PROPERTIES
-    OUTPUT_NAME ${exe_binary_name})
-
-target_include_directories(${exe_target_name}
-    PRIVATE ${CMAKE_SOURCE_DIR}/projects)
-
-target_link_libraries(${exe_target_name} PRIVATE
-    ${lib_target_name}
-    ores.risk.lib
-    ores.utility.lib
-    Boost::boost
-    Boost::date_time
-    Boost::program_options
-    ${CMAKE_THREAD_LIBS_INIT})
-
-
-install(TARGETS ${exe_target_name} RUNTIME DESTINATION bin
-    COMPONENT applications)
