+++ projects/ores.risk/ores.risk.hpp
@@ -1,6 +1,6 @@
 /* -*- mode: c++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
  *
- * Copyright (C) 2024 Marco Craveiro <marco.craveiro@gmail.com>
+ * Copyright (C) 2025 Marco Craveiro <marco.craveiro@gmail.com>
  *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
@@ -9,23 +9,23 @@
  *
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
+ * FOR A PARTICULAR PURPOSE. Seethe GNU General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License along with
+ * You should have received a copy of the GNU General PublicLicense along with
  * this program; if not, write to the Free Software Foundation, Inc., 51
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  *
  */
-#include <vector>
-#include <sstream>
-#include <boost/test/unit_test.hpp>
-#include <boost/algorithm/string/predicate.hpp>
+#ifndef ORES_RISK_RISK_HPP
+#define ORES_RISK_RISK_HPP
 
-BOOST_AUTO_TEST_SUITE(test_suite)
+#if defined(_MSC_VER) && (_MSC_VER >= 1200)
+#pragma once
+#endif
 
-BOOST_AUTO_TEST_CASE(test) {
-    BOOST_CHECK(true);
-}
+/**
+ * @brief Domain model for the ORE types and other related classes.
+ */
+namespace ores::risk {}
 
-BOOST_AUTO_TEST_SUITE_END()
+#endif
