+++ projects/ores.risk/types/currency.cpp
@@ -1,6 +1,6 @@
 /* -*- mode: c++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
  *
- * Copyright (C) 2025 Marco Craveiro <marco.craveiro@gmail.com>
+ * Copyright (C) 2024 Marco Craveiro <marco.craveiro@gmail.com>
  *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
@@ -17,28 +17,16 @@
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  *
  */
-#ifndef ORES_CORE_RISK_TYPES_CURRENCY_CONFIG_HPP
-#define ORES_CORE_RISK_TYPES_CURRENCY_CONFIG_HPP
+#include <ostream>
+#include <rfl.hpp>
+#include <rfl/json.hpp>
+#include "ores.risk/types/currency.hpp"
 
-#if defined(_MSC_VER) && (_MSC_VER >= 1200)
-#pragma once
-#endif
+namespace ores::risk::types {
 
-#include <iosfwd>
-#include <vector>
-#include "ores.core/risk/types/currency.hpp"
-
-namespace ores::core::risk::types {
-
-/**
- * @brief Represents a set of ORE currencies.
- */
-struct currency_config {
-    std::vector<currency> currencies;
-};
-
-std::ostream& operator<<(std::ostream& s, const currency_config& v);
-
+std::ostream& operator<<(std::ostream& s, const currency& v) {
+    rfl::json::write(v, s);
+    return(s);
 }
 
-#endif
+}
